% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-subsets.R
\name{subset}
\alias{subset}
\alias{subset.ps}
\title{Subsets of probability space objects.}
\usage{
\method{subset}{ps}(x, subset, ...)
}
\arguments{
\item{x}{a probability space}

\item{subset}{logical expression indicating elements or rows of space to keep: missing values are taken as false.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Subsets of probability space objects.
}
