% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces-examples.r
\name{tosscoin}
\alias{tosscoin}
\title{Tossing a coin}
\usage{
tosscoin(times, makespace = FALSE)
}
\arguments{
\item{times}{number of tosses.}

\item{makespace}{logical.}
}
\value{
A data frame, with an equally likely \code{probs} column if \code{makespace} is \code{TRUE}.
}
\description{
Sets up a sample space for the experiment of tossing a coin repeatedly with the outcomes "H" or "T".
}
\details{
The function uses `expand.grid()` to generate all possible sequences of flips resulting from 
the experiment of tossing a coin.  Columns of the dataframe are denoted `toss1`, 
`toss2`, up to `tosstimes`.
}
\examples{
tosscoin(2)
tosscoin(3, makespace = TRUE)
}
