% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces-examples.r
\name{urnsamples}
\alias{urnsamples}
\title{Sampling from urns}
\usage{
urnsamples(x, ...)
}
\arguments{
\item{x}{a vector or data frame from which sampling should take place.}

\item{...}{further arguments to be passed to or from other methods.}

\item{size}{number indicating the sample size.}

\item{replace}{logical indicating whether sampling should be done with replacement.}

\item{ordered}{logical indicating whether order among samples is important.}
}
\value{
A data frame if `urn` is a vector, and a list if `urn` is a data frame.
}
\description{
This function creates a sample space associated with the experiment of sampling 
distinguishable objects from an urn.
}
\details{
The function operates on the indices of the urn (or rows, in the case `urn` 
is a data frame).  It then takes those samples and substitutes back into `urn` 
to generate the entries of the data frame (or list, respectively).  In the 
case that `urn` has repeated values, the result will be repeated values 
in the output.

Note that `urnsamples` strips `x` of any existing `probs` 
column before sampling.
}
\examples{
urnsamples(1:10, size = 5)
S <- cards()
urnsamples(S, size = 2)
}
