% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-package.R
\docType{package}
\name{prob-package}
\alias{prob}
\alias{prob-package}
\title{prob: Elementary Probability on Finite Sample Spaces}
\description{
A framework for performing elementary probability calculations on finite sample spaces, which may be represented by data frames or lists. Functionality includes setting up sample spaces, counting tools, defining probability spaces, performing set algebra, calculating probability and conditional probability, tools for simulation and checking the law of large numbers, adding random variables, and finding marginal distributions. Characteristic functions for most base R distributions are implemented from the work of Eugene Lukacs (1972) \doi{10.2307/1425805}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gjkerns/prob/}
}

}
\author{
\strong{Maintainer}: G. Jay Kerns \email{gkerns@ysu.edu} [copyright holder]

}
\keyword{internal}
